Changed input of higher order parameters
In the initialization of the dynamics_Boris_multipole file, the
B_multip is dvivided by the factorial of the order.
This makes a better interface of the code.
Now, the user can specify the field derivatives instead of the fields
at x, y = 1, 0.
Also, B_multip in the c code is renamed to b_field, and B_skew is
renamed to B_field_skew